+++ devtools/projects/shell-browser/src/app/ng-validate.ts
@@ -8,68 +8,13 @@
 
 /// <reference types="chrome"/>
 
-export interface AngularDetection {
-  // This is necessary because the runtime
-  // message listener handles messages globally
-  // including from other extensions. We don't
-  // want to set icon and/or popup based on
-  // a message coming from an unrelated extension.
-  isAngularDevTools: true;
-  isIvy: boolean;
-  isAngular: boolean;
-  isDebugMode: boolean;
-  isSupportedAngularVersion: boolean;
-}
-
 window.addEventListener('message', (event: MessageEvent) => {
   if (event.source === window && event.data) {
     chrome.runtime.sendMessage(event.data);
   }
 });
 
-function detectAngular(win: Window): void {
-  const isDebugMode = Boolean((win as any).ng);
-  const ngVersionElement = document.querySelector('[ng-version]');
-  let isSupportedAngularVersion = false;
-  let isAngular = false;
-  if (ngVersionElement) {
-    isAngular = true;
-    const attr = ngVersionElement.getAttribute('ng-version');
-    const major = attr ? parseInt(attr.split('.')[0], 10) : -1;
-    // In case of g3 apps we support major 0.
-    if (attr && (major >= 9 || major === 0)) {
-      isSupportedAngularVersion = true;
-    }
-  }
-
-  win.postMessage(
-      {
-        // Needs to be inline because we're stringifying
-        // this function and executing it with eval.
-        isIvy: typeof (ngVersionElement as any)?.__ngContext__ !== 'undefined',
-        isAngular,
-        isDebugMode,
-        isSupportedAngularVersion,
-        isAngularDevTools: true,
-      } as AngularDetection,
-      '*');
-
-  if (!isAngular) {
-    setTimeout(() => detectAngular(win), 1000);
-  }
-}
-
-function installScript(fn: string): void {
-  const source = `;(${fn})(window)`;
-  const script = document.createElement('script');
-  script.textContent = source;
-  document.documentElement.appendChild(script);
-  const parentElement = script.parentElement;
-  if (parentElement) {
-    parentElement.removeChild(script);
-  }
-}
-
-if (document instanceof Document) {
-  installScript(detectAngular.toString());
-}
+const script = document.createElement('script');
+script.src = chrome.runtime.getURL('app/detect_angular_for_extension_icon_bundle.js');
+document.documentElement.appendChild(script);
+document.documentElement.removeChild(script);
