+++ devtools/projects/shell-browser/src/manifest/manifest.chrome.json
@@ -1,36 +1,48 @@
 {
-  "manifest_version": 2,
+  "manifest_version": 3,
   "short_name": "Angular DevTools",
   "name": "Angular DevTools",
   "description": "Angular DevTools extends Chrome DevTools adding Angular specific debugging and profiling capabilities.",
   "version": "1.0.6",
   "version_name": "1.0.6",
-  "minimum_chrome_version": "50",
-  "content_security_policy": "script-src 'self' 'unsafe-eval'; object-src 'self'",
+  "minimum_chrome_version": "102",
+  "content_security_policy": {
+    "extension_pages": "script-src 'self'; object-src 'self'"
+  },
   "icons": {
     "16": "assets/icon16.png",
     "48": "assets/icon48.png",
     "128": "assets/icon128.png"
   },
-  "browser_action": {
-    "default_popup": "popups/not-angular.html"
+  "action": {
+    "default_popup": "popups/not-angular.html",
+    "default_icon": {
+      "16": "assets/icon-bw16.png",
+      "48": "assets/icon-bw48.png",
+      "128": "assets/icon-bw128.png"
+    }
   },
   "devtools_page": "devtools.html",
   "web_accessible_resources": [
-    "app/backend_bundle.js",
-    "devtools.html"
+    {
+      "resources": [
+        "app/backend_bundle.js",
+        "app/detect_angular_for_extension_icon_bundle.js",
+        "devtools.html"
+      ],
+      "matches": ["<all_urls>"],
+      "extension_ids": []
+    }
   ],
   "background": {
-    "scripts": [
-      "app/background_bundle.js"
-    ],
-    "persistent": false
+    "service_worker": "app/background_bundle.js"
   },
   "permissions": [
-    "activeTab",
-    "http://*/*",
-    "https://*/*",
-    "file:///*"
+    "scripting",
+    "activeTab"
+  ],
+  "host_permissions": [
+    "<all_urls>"
   ],
   "content_scripts": [
     {
